+++ test/fixture/unused/condition-true-literal-expression.js
@@ -1,0 +1,4 @@
+const A = true ? { } : { a };
+const A2 = true || x || 2 || '2' || fn() ? { } : { a2 };
+const A3 = x || 2 || false || true || '2' || fn() ? { } : { a3 };
+const A4 = x || 2 || false || true || '2' || fn() ? { } : { a3 }, A41 = {};