+++ test/fixture/unused/if-false-literal-statement.js
@@ -1,0 +1,4 @@
+if (false) { a } else {}
+if (false && x && 2 && true && fn()) { a2 } else {}
+if (x && 2 && true && '2' && false) { a3 } else {}
+if (x && 2 && true) {} else if ('2' && false) { a3 } else {}