+++ test/fixture/unused/if-false-statement.js
@@ -1,0 +1,5 @@
+const isX = false;
+if (isX) { a } else {}
+if (isX && x && 2 && true && fn()) { a2 } else {}
+if (x && 2 && true && '2' && isX) { a3 } else {}
+if (x && 2 && true) {} else if ('2' && isX) { a41 } else {}