+++ test/fixture/unused/if-true-literal-statement.js
@@ -1,0 +1,5 @@
+if (true) { } else { a }
+if (true || x || 2 || false || fn()) { } else { a2 }
+if (x || 2 || false || '2' || true) { } else { a3 }
+if (false) { a4 } else if (true) { } else { a41 }
+if (false) { a5 } else if (x || 2 || true || false || '2') { } else { a51 }