+++ test/fixture/unused/if-true-statement.js
@@ -1,0 +1,21 @@
+const isX = true;
+if (isX) {
+} else {
+  a;
+}
+if (isX || x || 2 || false || fn()) {
+} else {
+  a2;
+}
+if (x || 2 || false || "2" || isX) {
+} else {
+  a3;
+}
+if (false) {
+  a4;
+} else if (true) {
+} else if (isWeb) {
+  a41;
+} else {
+  a42;
+}