+++ test/fixture/used/if-slide-effects-false-literal-statement.js
@@ -1,0 +1,1 @@
+if (fn() && false) { a0 } else {}