+++ test/fixture/used/if-slide-effects-false-statement.js
@@ -1,0 +1,2 @@
+const isX = false;
+if (fn() && isX) { a0 } else {}