+++ test/fixture/used/if-slide-effects-true-statement.js
@@ -1,0 +1,2 @@
+const isX = true;
+if (fn() || isX) { } else { a0 }