+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.3 (2024-11-11)
+
+- Fix: compat unused fix-case1
+- Test: add unit tests
+- Chore: add babel transform, target node16
+
 ## 1.0.2 (2024-10-31)
 
 - Feat: support remove condition and if of booleanLiteral `true` and `const isX = true`